back/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
axios==0.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.7
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==2.2.5
Django==4.2.16
django-allauth==0.54.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.3
oauthlib==3.2.2
openai==1.55.0
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
vue==0.0.1
websocket-client==1.8.0
wsproto==1.2.0
xmltodict==0.14.2
